added alert delete dialog

+++ src/components/Noteitem.jsx
@@ -4,6 +4,17 @@ import { FaEdit } from "react-icons/fa";
 import noteContext from "../context/notes/NoteContext";
 import { useContext } from "react";
 import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+import {
   Card,
   CardContent,
   CardFooter,
@@ -30,13 +41,32 @@ const Noteitem = (props) => {
           {note.tag && <p className="text-sm text-gray-400">#{note.tag}</p>}
 
           <div className="w-full flex justify-end">
-            <ImBin2
-              className="delbtn mr-2"
-              onClick={() => {
-                deleteNote(note._id);
-                toast.success("Note Deleted");
-              }}
-            />
+            <AlertDialog>
+              <AlertDialogTrigger asChild>
+                <ImBin2 className="delbtn mr-2" />
+              </AlertDialogTrigger>
+              <AlertDialogContent>
+                <AlertDialogHeader>
+                  <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+                  <AlertDialogDescription>
+                    This action cannot be undone. This will permanently delete
+                    your note and remove your data from our database.
+                  </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter>
+                  <AlertDialogCancel>Cancel</AlertDialogCancel>
+                  <AlertDialogAction
+                    onClick={() => {
+                      deleteNote(note._id);
+                      toast.success("Note Deleted");
+                    }}
+                  >
+                    Delete
+                  </AlertDialogAction>
+                </AlertDialogFooter>
+              </AlertDialogContent>
+            </AlertDialog>
+
             <FaEdit
               className="editbtn "
               onClick={() => {
